refactor: 하드코딩한 숫자 상수로 변경

+++ src/main/java/lotto/domain/Prize.java
@@ -15,7 +15,7 @@ public enum Prize {
         this.expression = expression;
     }
 
-    public int calculate(Integer value) {
+    public double calculate(Integer value) {
         return expression.apply(value);
     }
 }
